created User services/interface and implemented changes in controllers, blog services, and ninject
