+++ Blog/GamingGuruBlog.Web/GamingGuruBlog.Domain/GamingGuruBlog.Domain.csproj
@@ -70,9 +70,10 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CategoryServices.cs" />
+    <Compile Include="Interfaces\IUserServices.cs" />
+    <Compile Include="Services\CategoryServices.cs" />
     <Compile Include="Interfaces\IBlogServices.cs" />
-    <Compile Include="BlogServices.cs" />
+    <Compile Include="Services\BlogServices.cs" />
     <Compile Include="Interfaces\IBlogCategoryRepository.cs" />
     <Compile Include="Interfaces\IBlogPostRepository.cs" />
     <Compile Include="Interfaces\IBlogTagRepository.cs" />
@@ -93,7 +94,8 @@
     <Compile Include="Models\Tag.cs" />
     <Compile Include="Models\User.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="StaticPageServices.cs" />
+    <Compile Include="Services\StaticPageServices.cs" />
+    <Compile Include="Services\UserServices.cs" />
     <Compile Include="TagServices.cs" />
   </ItemGroup>
   <ItemGroup>
