+++ Blog/GamingGuruBlog.Web/GamingGuruBlog.Domain/Interfaces/IBlogServices.cs
@@ -20,9 +20,7 @@ namespace GamingGuruBlog.Domain.Interfaces
         void DeleteBlogPost(int blogID);
 
         void AddNewBlogPost(BlogPost newPost);
+
         void ProcessEditedBlogPost(BlogPost editedBlogPost);
-        User GetUser(string userID);
-        void EditUser(User editedUser);
-        List<User> GetAllUsers();
     }
 }
