+++ Blog/GamingGuruBlog.Web/GamingGuruBlog.Domain/Interfaces/IUserServices.cs
@@ -1,0 +1,16 @@
+﻿using GamingGuruBlog.Domain.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GamingGuruBlog.Domain.Interfaces
+{
+    public interface IUserServices
+    {
+        User GetUser(string userID);
+        void EditUser(User editedUser);
+        List<User> GetAllUsers();
+    }
+}
