+++ Blog/GamingGuruBlog.Web/GamingGuruBlog.Web/App_Start/NinjectWebCommon.cs
@@ -11,6 +11,7 @@ namespace GamingGuruBlog.Web.App_Start
     using Domain.Interfaces;
     using Data.Repositories;
     using Domain;
+    using Domain.Services;
 
     public static class NinjectWebCommon 
     {
@@ -72,6 +73,7 @@ namespace GamingGuruBlog.Web.App_Start
             kernel.Bind<IBlogServices>().To<BlogServices>();
             kernel.Bind<IStaticPageServices>().To<StaticPageServices>();
             kernel.Bind<ICategoryServices>().To<CategoryServices>();
+            kernel.Bind<IUserServices>().To<UserServices>();
         }
     }
 }
