+++ Blog/GamingGuruBlog.Web/GamingGuruBlog.Web/Controllers/PartialViewController.cs
@@ -12,14 +12,16 @@ namespace GamingGuruBlog.Web.Controllers
         private IStaticPageServices _staticPageServices;
         private ICategoryServices _categoryServices;
         private ITagServices _tagServices;
+        private IUserServices _userServices;
 
 
-        public PartialViewController(IBlogServices newBlogservices, IStaticPageServices newStaticPageServices, ICategoryServices newCategoryServices, ITagServices newTagServices)
+        public PartialViewController(IBlogServices newBlogservices, IStaticPageServices newStaticPageServices, ICategoryServices newCategoryServices, ITagServices newTagServices, IUserServices newUserServices)
         {
             _blogServices = newBlogservices;
             _staticPageServices = newStaticPageServices;
             _categoryServices = newCategoryServices;
             _tagServices = newTagServices;
+            _userServices = newUserServices;
         }
         // GET: PartialView
         public PartialViewResult Action()
@@ -40,7 +42,7 @@ namespace GamingGuruBlog.Web.Controllers
 
         public PartialViewResult AdminPanelTabbedUsers()
         {
-            var model = _blogServices.GetAllUsers();
+            var model = _userServices.GetAllUsers();
             return PartialView("~/Views/PartialView/AdminPanelTabbedUsers.cshtml", model);
         }
